requirements.txt
langchain
langchain-community
langchain-google-genai
python-dotenv
Flask
Flask-Cors
pillow